backend/requirements.txt
fastapi
uvicorn
python-multipart
PyPDF2
python-docx
spacy
sentence-transformers
torch
celery
redis

